+++ bin/baseConfig.js
@@ -80,7 +80,7 @@ module.exports = {
     },
     jobs: {
         topicConfigPollingMs: 15000,
-        metadataFetcherMs: 23 * 62000,
+        metadataFetcherMs: 39 * 60 * 1000,
     },
     hooks: {
         enabled: true,
@@ -88,5 +88,7 @@ module.exports = {
         only: false,
         retries: 2,
         retryTimeoutMs: 1000,
+        subscriptionConcurrency: 4,
+        replayConcurrency: 4,
     },
 };