+++ lib/api/routes/replay.ts
@@ -4,24 +4,45 @@ import Zamza from "../../Zamza";
 const routeReplay = (zamza: Zamza) => {
 
     const router = express.Router();
-    const replayModel = zamza.mongoWrapper.getReplay();
     const replayHandler = zamza.replayHandler;
 
     // root api GET is in topic-config.ts
 
     router.get("/replay", async (req, res) => {
         try {
-            const replay = await replayModel.get();
-
-            if (replay) {
-                res.status(200).json(replay);
-                return;
-            }
-
-            res.status(404).json({
-                error: "Replay does not exist.",
+            res.status(200).json(await replayHandler.getCurrentReplay());
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
             });
+        }
+    });
 
+    router.get("/replays", async (req, res) => {
+        try {
+            res.status(200).json(await replayHandler.listReplays());
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
+            });
+        }
+    });
+
+    router.get("/replay/lag", async (req, res) => {
+        try {
+            res.status(200).json(replayHandler.mirrorConsumer ?
+                await replayHandler.mirrorConsumer.getLagStatus() : null);
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
+            });
+        }
+    });
+
+    router.get("/replay/analytics", async (req, res) => {
+        try {
+            res.status(200).json(replayHandler.mirrorConsumer ?
+                await replayHandler.mirrorConsumer.getAnalytics() : null);
         } catch (error) {
             res.status(500).json({
                 error: "An error occured " + error.message,
@@ -41,17 +62,25 @@ const routeReplay = (zamza: Zamza) => {
         const { topic, consumerGroup } = req.body || ({} as any);
         if (!topic) {
             res.status(400).json({
-                error: "Body should be a valid object, {topic}",
+                error: "Body should be a valid object, {topic, consumerGroup?}",
+            });
+            return;
+        }
+
+        if (replayHandler.isCurrentlyRunning()) {
+            res.status(400).json({
+                error: "Replay configuration active, can only have one replay" +
+                    "at a time, delete currently running one first.",
             });
             return;
         }
 
         try {
-            const existingReplay = await replayModel.get();
+            const existingReplay = await replayHandler.isBeingReplayedByAnyInstance(topic);
             if (existingReplay) {
                 res.status(400).json({
                     error: "Replay configuration active, can only have one replay" +
-                        "at a time, delete currently running one first.",
+                        " per topic at a time, delete currently running one first.",
                 });
                 return;
             }
@@ -65,7 +94,34 @@ const routeReplay = (zamza: Zamza) => {
         }
     });
 
-    router.delete("/replay", async (req, res) => {
+    router.delete("/replay/:topic", async (req, res) => {
+
+        if (!res.locals.access.replayAccessAllowedForRequest(req)) {
+            res.status(403).json({
+                error: "Access not allowed",
+            });
+            return;
+        }
+
+        if (!replayHandler.dealsWithTopic(req.params.topic)) {
+            res.status(400).json({
+                error: "This instance does not deal with the provided topic. It deals with: " +
+                    replayHandler.currentTargetTopic,
+            });
+            return;
+        }
+
+        try {
+            await replayHandler.stopReplay();
+            res.status(204).end();
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
+            });
+        }
+    });
+
+    router.delete("/replay/flushall", async (req, res) => {
 
         if (!res.locals.access.replayAccessAllowedForRequest(req)) {
             res.status(403).json({
@@ -75,7 +131,26 @@ const routeReplay = (zamza: Zamza) => {
         }
 
         try {
-            await replayHandler.stopReplay();
+            await replayHandler.flushall();
+            res.status(204).end();
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
+            });
+        }
+    });
+
+    router.delete("/replay/flushone", async (req, res) => {
+
+        if (!res.locals.access.replayAccessAllowedForRequest(req)) {
+            res.status(403).json({
+                error: "Access not allowed",
+            });
+            return;
+        }
+
+        try {
+            await replayHandler.flushone();
             res.status(204).end();
         } catch (error) {
             res.status(500).json({