+++ lib/interfaces/Replay.ts
@@ -1,4 +1,5 @@
 export interface Replay {
+    instanceId: string;
     topic: string;
     timestamp: number;
     consumerGroup: string;